Add validator.py
