+++ maps/test_5.json
@@ -14,7 +14,7 @@
          "y":0
         }, 
         {
-         "data":[30, 31, 32, 33, 34, 35, 36, 37, 31, 24, 18, 1610612760, 40, 41, 42, 43, 44, 45, 46, 47, 4, 13, 13, 19, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 20, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 21, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 22, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 23, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 1, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 25, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 24, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 18, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 11, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10],
+         "data":[30, 31, 32, 33, 34, 35, 38, 39, 0, 0, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 0, 0, 0, 0, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 25, 26, 27, 28, 29, 24, 18, 23, 22, 21, 20, 19, 25, 26, 27, 28, 29, 24, 18, 23, 22, 21, 20, 19, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10, 2, 3, 4, 16, 17, 14, 15, 8, 7, 38, 39, 10],
          "height":12,
          "id":2,
          "name":"layer 2",
@@ -26,7 +26,7 @@
          "y":0
         }, 
         {
-         "data":[13, 0, 0, 0, 0, 0, 0, 0, 31, 13, 13, 18, 13, 0, 0, 0, 0, 0, 0, 0, 4, 18, 24, 25, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 26, 2, 3, 3221225476, 16, 17, 14, 15, 8, 7, 40, 41, 27, 2, 3, 4, 16, 17, 14, 15, 8, 7, 30, 31, 28, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 1610612754, 29, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 1610612760, 0, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612750, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612752, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 2684354564, 1610612739, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2],
+         "data":[0, 0, 0, 0, 0, 0, 36, 37, 0, 0, 10, 11, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18, 24, 7, 8, 10, 11, 3, 4, 16, 17, 14, 15, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 13, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612750, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612751, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612753, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612752, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612739, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740, 1610612740],
          "height":12,
          "id":3,
          "name":"layer 3",
@@ -38,7 +38,7 @@
          "y":0
         }, 
         {
-         "data":[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+         "data":[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
          "height":12,
          "id":5,
          "name":"layer 4",
